core/transfer: moved to core/wallet/transfer
